Rename class to `InfrahubGraphQLQueryAnalyzer`
Move some GraphQL analyzer tests to the SDK tests.
