+++ backend/infrahub/graphql/analyzer.py
@@ -4,7 +4,7 @@ from graphql import (
     GraphQLSchema,
     OperationType,
 )
-from infrahub_sdk.analyzer import GraphQLQueryAnalyzerBase
+from infrahub_sdk.analyzer import GraphQLQueryAnalyzer
 from infrahub_sdk.utils import extract_fields
 
 from infrahub.core import registry
@@ -12,7 +12,7 @@ from infrahub.core.branch import Branch
 from infrahub.graphql.utils import extract_schema_models
 
 
-class GraphQLQueryAnalyzer(GraphQLQueryAnalyzerBase):
+class InfrahubGraphQLQueryAnalyzer(GraphQLQueryAnalyzer):
     def __init__(self, query: str, schema: Optional[GraphQLSchema] = None, branch: Optional[Branch] = None):
         self.branch: Optional[Branch] = branch
         super().__init__(query=query, schema=schema)
