+++ backend/infrahub/graphql/auth/query_permission_checker/anonymous_checker.py
@@ -2,7 +2,7 @@ from typing import Callable
 
 from infrahub.auth import AccountSession
 from infrahub.exceptions import AuthorizationError
-from infrahub.graphql.analyzer import GraphQLQueryAnalyzer
+from infrahub.graphql.analyzer import InfrahubGraphQLQueryAnalyzer
 
 from .interface import GraphQLQueryPermissionCheckerInterface
 
@@ -14,7 +14,7 @@ class AnonymousGraphQLPermissionChecker(GraphQLQueryPermissionCheckerInterface):
     async def supports(self, account_session: AccountSession) -> bool:
         return not account_session.authenticated
 
-    async def check(self, analyzed_query: GraphQLQueryAnalyzer):
+    async def check(self, analyzed_query: InfrahubGraphQLQueryAnalyzer):
         if self.anonymous_access_allowed_func() and not analyzed_query.contains_mutation:
             return
         raise AuthorizationError("Authentication is required to perform this operation")
