+++ backend/infrahub/graphql/auth/query_permission_checker/checker.py
@@ -2,7 +2,7 @@ from typing import List
 
 from infrahub.auth import AccountSession
 from infrahub.exceptions import PermissionDeniedError
-from infrahub.graphql.analyzer import GraphQLQueryAnalyzer
+from infrahub.graphql.analyzer import InfrahubGraphQLQueryAnalyzer
 
 from .interface import GraphQLQueryPermissionCheckerInterface
 
@@ -11,7 +11,7 @@ class GraphQLQueryPermissionChecker:
     def __init__(self, sub_checkers: List[GraphQLQueryPermissionCheckerInterface]):
         self.sub_checkers = sub_checkers
 
-    async def check(self, account_session: AccountSession, analyzed_query: GraphQLQueryAnalyzer):
+    async def check(self, account_session: AccountSession, analyzed_query: InfrahubGraphQLQueryAnalyzer):
         for sub_checker in self.sub_checkers:
             if await sub_checker.supports(account_session):
                 await sub_checker.check(analyzed_query)
