+++ backend/infrahub/graphql/auth/query_permission_checker/default_checker.py
@@ -1,6 +1,6 @@
 from infrahub.auth import AccountSession
 from infrahub.exceptions import AuthorizationError
-from infrahub.graphql.analyzer import GraphQLQueryAnalyzer
+from infrahub.graphql.analyzer import InfrahubGraphQLQueryAnalyzer
 
 from .interface import GraphQLQueryPermissionCheckerInterface
 
@@ -9,5 +9,5 @@ class DefaultGraphQLPermissionChecker(GraphQLQueryPermissionCheckerInterface):
     async def supports(self, account_session: AccountSession) -> bool:
         return True
 
-    async def check(self, analyzed_query: GraphQLQueryAnalyzer):
+    async def check(self, analyzed_query: InfrahubGraphQLQueryAnalyzer):
         raise AuthorizationError("Authentication is required to perform this operation")
