+++ backend/infrahub/graphql/auth/query_permission_checker/read_write_checker.py
@@ -1,5 +1,5 @@
 from infrahub.auth import AccountSession
-from infrahub.graphql.analyzer import GraphQLQueryAnalyzer
+from infrahub.graphql.analyzer import InfrahubGraphQLQueryAnalyzer
 
 from .interface import GraphQLQueryPermissionCheckerInterface
 
@@ -8,5 +8,5 @@ class ReadWriteGraphQLPermissionChecker(GraphQLQueryPermissionCheckerInterface):
     async def supports(self, account_session: AccountSession) -> bool:
         return account_session.authenticated and not account_session.read_only
 
-    async def check(self, analyzed_query: GraphQLQueryAnalyzer):
+    async def check(self, analyzed_query: InfrahubGraphQLQueryAnalyzer):
         return
