+++ backend/tests/unit/graphql/auth/test_default_checker.py
@@ -5,14 +5,14 @@ import pytest
 from infrahub.auth import AccountSession, AuthType
 from infrahub.core.constants import AccountRole
 from infrahub.exceptions import AuthorizationError
-from infrahub.graphql.analyzer import GraphQLQueryAnalyzer
+from infrahub.graphql.analyzer import InfrahubGraphQLQueryAnalyzer
 from infrahub.graphql.auth.query_permission_checker.default_checker import DefaultGraphQLPermissionChecker
 
 
 class TestDefaultAuthChecker:
     def setup_method(self):
         self.account_session = AccountSession(account_id="abc", auth_type=AuthType.JWT)
-        self.graphql_query = AsyncMock(spec=GraphQLQueryAnalyzer)
+        self.graphql_query = AsyncMock(spec=InfrahubGraphQLQueryAnalyzer)
         self.checker = DefaultGraphQLPermissionChecker()
 
     @pytest.mark.parametrize("role", [x.value for x in AccountRole])
