+++ backend/tests/unit/graphql/auth/test_parent_checker.py
@@ -5,7 +5,7 @@ import pytest
 from infrahub.auth import AccountSession, AuthType
 from infrahub.core.constants import AccountRole
 from infrahub.exceptions import PermissionDeniedError
-from infrahub.graphql.analyzer import GraphQLQueryAnalyzer
+from infrahub.graphql.analyzer import InfrahubGraphQLQueryAnalyzer
 from infrahub.graphql.auth.query_permission_checker.checker import GraphQLQueryPermissionChecker
 from infrahub.graphql.auth.query_permission_checker.interface import GraphQLQueryPermissionCheckerInterface
 
@@ -13,7 +13,7 @@ from infrahub.graphql.auth.query_permission_checker.interface import GraphQLQuer
 class TestParentAuthChecker:
     def setup_method(self):
         self.account_session = AccountSession(account_id="abc", auth_type=AuthType.JWT, role=AccountRole.ADMIN)
-        self.graphql_query = AsyncMock(spec=GraphQLQueryAnalyzer)
+        self.graphql_query = AsyncMock(spec=InfrahubGraphQLQueryAnalyzer)
         self.sub_auth_checker_one = AsyncMock(spec=GraphQLQueryPermissionCheckerInterface)
         self.sub_auth_checker_two = AsyncMock(spec=GraphQLQueryPermissionCheckerInterface)
         self.sub_auth_checker_one.supports.return_value = False
