+++ python_sdk/infrahub_sdk/__init__.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from infrahub_sdk.analyzer import GraphQLOperation, GraphQLQueryAnalyzerBase, GraphQLQueryVariable
+from infrahub_sdk.analyzer import GraphQLOperation, GraphQLQueryAnalyzer, GraphQLQueryVariable
 from infrahub_sdk.batch import InfrahubBatch
 from infrahub_sdk.branch import InfrahubBranchManager, InfrahubBranchManagerSync
 from infrahub_sdk.client import InfrahubClient, InfrahubClientSync
@@ -47,7 +47,7 @@ __all__ = [
     "InfrahubSchema",
     "FilterNotFound",
     "generate_uuid",
-    "GraphQLQueryAnalyzerBase",
+    "GraphQLQueryAnalyzer",
     "GraphQLQueryVariable",
     "GraphQLError",
     "GraphQLOperation",
