+++ python_sdk/tests/unit/sdk/test_utils.py
@@ -1,6 +1,7 @@
 import uuid
 
 import pytest
+from graphql import parse
 
 from infrahub_sdk.node import InfrahubNode
 from infrahub_sdk.utils import (
@@ -12,6 +13,7 @@ from infrahub_sdk.utils import (
     deep_merge_dict,
     dict_hash,
     duplicates,
+    extract_fields,
     get_flat_value,
     is_valid_url,
     is_valid_uuid,
@@ -130,3 +132,46 @@ def test_dict_hash():
     assert dict_hash({"b": 2, "a": {"c": 1, "d": 2}}) == "729f4b898271d3fa95a7363bdd7c215d"
     assert dict_hash({"b": 2, "a": {"d": 2, "c": 1}}) == "729f4b898271d3fa95a7363bdd7c215d"
     assert dict_hash({}) == "99914b932bd37a50b983c5e7c90ae93b"
+
+
+async def test_extract_fields(query_01):
+    document = parse(query_01)
+    expected_response = {
+        "TestPerson": {
+            "edges": {
+                "node": {
+                    "cars": {"edges": {"node": {"name": {"value": None}}}},
+                    "name": {"value": None},
+                },
+            },
+        },
+    }
+    assert await extract_fields(document.definitions[0].selection_set) == expected_response
+
+
+async def test_extract_fields_fragment(query_02):
+    document = parse(query_02)
+
+    expected_response = {
+        "TestPerson": {
+            "edges": {
+                "node": {
+                    "cars": {
+                        "edges": {
+                            "node": {
+                                "member_of_groups": {
+                                    "edges": {"node": {"id": None}},
+                                },
+                                "mpg": {"is_protected": None, "value": None},
+                                "name": {"value": None},
+                                "nbr_engine": {"value": None},
+                            },
+                        },
+                    },
+                    "name": {"value": None},
+                },
+            },
+        },
+    }
+
+    assert await extract_fields(document.definitions[0].selection_set) == expected_response
